+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Release 1.7.4
 	- added optional parameter dqpar for psd_psq or ld_lq (default) identification in pm parident
+    - added method transient for PMRelMachinePsidq (experimental)
 
 ## Release 1.7.0
 	- displ_stator_rotor: eccentricity simulation and analysis